AOSP: libext2fs: Track the libsparse API change.
libsparse is updating the 'len' parameter, from 'int' to 'size_t', in
the callback parameter of sparse_file_foreach_chunk(). The value
represents the chunk size, which could be legitimately larger than
INT_MAX.


Google-Bug-Id: 78432315
Test: `m dist` with aosp_marlin-userdebug
From AOSP commit: 019bed178935585d28cf702a8090a5598415312b

+++ lib/ext2fs/sparse_io.c
@@ -67,7 +67,7 @@ static void free_sparse_blocks(struct sparse_map *sm)
 	sm->blocks = NULL;
 }
 
-static int sparse_import_segment(void *priv, const void *data, int len,
+static int sparse_import_segment(void *priv, const void *data, size_t len,
 				 unsigned int block, unsigned int nr_blocks)
 {
 	struct sparse_map *sm = priv;
